pyproject.toml
[tool.poetry]
name = "reclamosapi"
version = "0.1.0"
description = "Backend para la aplicación de Reclamos de Apromed S.A.S."
authors = ["dfcarrera79 <44879886+dfcarrera79@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.30"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
